refactor: add tui feature
