+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod selectors;
 pub mod source;
 pub mod system_tools;
 pub mod tool_configuration;
+#[cfg(feature = "tui")]
 pub mod tui;
 pub mod used_variables;
 pub mod utils;
